backend/requirements.txt
Flask==2.0.1
numpy==1.21.2
midiutil==1.2.1
Pillow==8.3.1